feat: InputView, ResultView로 입출력 기능 분리
